+++ MatrixCounter/models/matrixmodel1.h
@@ -11,7 +11,7 @@ public:
     QVariant data (const QModelIndex& index, int role) const override;
     int data (int i, int j) const;
     bool setData (const QModelIndex& index, const QVariant& data, int role) override;
-    void setData (int i, int j, const int& data);
+    void setData (int i, int j, const double& data);
     void clearData();
 
     int rowCount (const QModelIndex& index = QModelIndex()) const override;
@@ -23,9 +23,9 @@ public:
     void setRowCount (int);
     void setColumnCount (int);
 
-    QVector<QVector<int>> GetVectoredMat();
+    QVector<QVector<double>> GetVectoredMat();
 private:
-    QHash<QModelIndex, int> mat;
+    QHash<QModelIndex, double> mat;
     int rows;
     int cols;
 };
